final_project/backend/requirements.txt
flask
flask-sqlalchemy
flask-login
flask-cors
python-dotenv
google-genai
werkzeug
reportlab
appwrite
